+++ emerlin2caom2/main_app.py
@@ -71,59 +71,91 @@ This module implements the ObsBlueprint mapping, as well as the workflow
 entry point that executes the workflow.
 """
 
-from os.path import basename
+import os
+import subprocess
 
-from caom2pipe import caom_composable as cc
-from caom2pipe import manage_composable as mc
+from caom2 import SimpleObservation, ObservationIntentType, Target, Telescope, TypedOrderedDict, Plane, Artifact, \
+    ReleaseType, ObservationWriter, ProductType, ChecksumURI
 
+import casa_reader as casa
+import measurement_set_metadata as msmd
 
 __all__ = [
-    'EMerlinMapping',
-    'EMerlin',
+    'emerlin_main_app'
 ]
 
 
-class EMerlin(mc.StorageName):
-    """Naming rules:
-    - support mixed-case file name storage, and mixed-case obs id values
-    - support uncompressed files in storage
-    """
-
-    E_MERLIN_PATTERN = '*'
-
-    def __init__(self, entry):
-        super().__init__(file_name=basename(entry), source_names=[entry])
-
-    def is_valid(self):
-        return True
+def basename(name):
+    base_name = os.path.dirname(name).split('/')[-1]
+    return base_name
 
 
-class EMerlinMapping(cc.TelescopeMapping):
-    def __init__(self, storage_name, headers, clients, observable, observation, config):
-        super().__init__(storage_name, headers, clients, observable, observation, config)
+def create_observation(storage_name, xml_out_dir):
+    obs_id = basename(storage_name)
+    observation = SimpleObservation('collection', obs_id)
+    observation.obs_type = 'science'
+    observation.intent = ObservationIntentType.SCIENCE
 
-    def accumulate_blueprint(self, bp):
-        """Configure the telescope-specific ObsBlueprint at the CAOM model Observation level."""
-        self._logger.debug('Begin accumulate_bp.')
-        super().accumulate_blueprint(bp)
+    observation.target = Target('TBD')
+    # observation.target.position = TargetPosition(str(find_mssources(ms_file)), 'J2000')
+    observation.telescope = Telescope(casa.get_obs_name(storage_name)[0])
 
-        bp.set('Plane.calibrationLevel', 1)
-        bp.set('Plane.dataProductType', 'image')
-        bp.set('Artifact.productType', 'science')
-        bp.set('Artifact.releaseType', 'data')
+    observation.planes = TypedOrderedDict(Plane)
+    plane = Plane(obs_id)
+    observation.planes[obs_id] = plane
 
-        bp.configure_position_axes((1, 2))
-        bp.configure_time_axis(3)
-        bp.configure_energy_axis(4)
-        bp.configure_polarization_axis(5)
-        bp.configure_observable_axis(6)
-        self._logger.debug('Done accumulate_bp.')
+    plane.artifacts = TypedOrderedDict(Artifact)
+    artifact = Artifact('uri:foo/bar', ProductType.SCIENCE, ReleaseType.META)
+    plane.artifacts['uri:foo/bar'] = artifact
 
-    def update(self, file_info):
-        """Called to fill multiple CAOM model elements and/or attributes (an n:n relationship between TDM attributes 
-        and CAOM attributes).
-        """
-        return super().update(file_info)
+    meta_data = msmd.get_local_file_info(storage_name)
 
-    def _update_artifact(self, artifact):
-        pass
+    artifact.content_type = meta_data.file_type
+    artifact.content_length = meta_data.size
+    artifact.content_checksum = ChecksumURI('md5:{}'.format(meta_data.md5sum))
+
+    xml_output_name = xml_out_dir + obs_id + '.xml'
+
+    writer = ObservationWriter()
+    writer.write(observation, xml_output_name)
+
+    return xml_output_name, obs_id
+
+
+def upload_xml(xml_output_name, observation_id, rootca_cert, repo_url_base='https://src-data-repo.co.uk/torkeep/',
+               collection='EMERLIN'):
+
+    if rootca_cert is None:
+        put_command = ['curl', '-X', 'PUT', '-T', xml_output_name,
+                       repo_url_base+collection+'/'+observation_id]
+        post_command = ['curl', '-X', 'POST', '-T', xml_output_name,
+                        repo_url_base+collection+'/'+observation_id]
+    else:
+        put_command = ['curl', '-ca', '"'+rootca_cert+'"', '-X', 'PUT', '-T', xml_output_name,
+                       repo_url_base+collection+'/'+observation_id]
+        post_command = ['curl', '-ca', '"'+rootca_cert+'"', '-X', 'POST', '-T', xml_output_name,
+                        repo_url_base+collection+'/'+observation_id]
+    # print(put_command)
+    # print(post_command)
+    # subprocess.call(put_command)
+    try:
+        subprocess.call(put_command) # method 405, not allowed
+    except subprocess.CalledProcessError:
+        subprocess.call(post_command)
+
+
+def emerlin_main_app(storage_name, rootca=None, xml_dir='.'):
+
+    xml_output_file, obs_id = create_observation(storage_name, xml_dir)
+    upload_xml(xml_output_file, obs_id, rootca)
+
+    ## add something like this for logging later
+    # try:
+    #     result = to_caom2()
+    #     sys.exit(result)
+    # except Exception as e:
+    #     logging.error(
+    #         f'Failed {APPLICATION} execution for {args} with {str(e)}.')
+    #     tb = traceback.format_exc()
+    #     logging.error(tb)
+    #     sys.exit(-1)
