+++ emerlin2caom2/tests/test_composable.py
@@ -83,8 +83,8 @@ def test_run_by_state():
 def test_run(run_mock, access_mock, test_config, tmp_path):
     run_mock.return_value = 0
     access_mock.return_value = 'https://localhost'
-    test_f_id = 'test_file_id'
-    test_f_name = f'{test_f_id}.fits'
+    test_f_id = '/home/h14471mj/e-merlin/casa6_docker/prod/TS8004_C_001_20190801/TS8004_C_001_20190801_avg'
+    test_f_name = f'{test_f_id}.ms'
     orig_cwd = os.getcwd()
     try:
         os.chdir(tmp_path.as_posix())
