+++ emerlin2caom2/tests/test_storage_name.py
@@ -65,29 +65,27 @@
 #
 # ***********************************************************************
 #
-
-from caom2pipe import manage_composable as mc
-from emerlin2caom2 import BlankName
+from emerlin2caom2 import EMerlin
 
 
 def test_is_valid():
-    assert BlankName('anything').is_valid()
+    assert EMerlin('anything').is_valid()
     
 
 def test_storage_name(test_config):
-    test_obs_id = 'TEST_OBS_ID'
-    test_f_name = f'{test_obs_id}.fits'
+    test_obs_id = '/home/h14471mj/e-merlin/casa6_docker/prod/TS8004_C_001_20190801/TS8004_C_001_20190801_avg'
+    test_f_name = f'{test_obs_id}.ms'
     test_uri = f'{test_config.scheme}:{test_config.collection}/{test_f_name}'
-   for index, entry in enumerate(
-        [
-            test_f_name, 
-            test_uri, 
-            f'https://localhost:8020/{test_f_name}', 
-            f'vos:goliaths/test/{test_f_name}',
-            f'/tmp/{test_f_name}',
-        ]   
+    for index, entry in enumerate(
+         [
+             test_f_name,
+             test_uri,
+             f'https://localhost:8020/{test_f_name}',
+             f'vos:goliaths/test/{test_f_name}',
+             f'/tmp/{test_f_name}',
+         ]
     ):
-        test_subject = BlankName(entry)
+        test_subject = EMerlin(entry)
         assert test_subject.file_id == test_f_name.replace('.fits', '').replace('.header', ''), f'wrong file id {index}'
         assert test_subject.file_uri == test_uri, f'wrong uri {index}'
         assert test_subject.obs_id == test_obs_id, f'wrong obs id {index}'
