+++ scripts/rootCA.pem
@@ -1,0 +1,23 @@
+-----BEGIN CERTIFICATE-----
+MIID4zCCAsugAwIBAgIUUkMgDKh3lfso5iGX1Od5SXL8v8AwDQYJKoZIhvcNAQEL
+BQAwgYAxCzAJBgNVBAYTAlVLMRIwEAYDVQQIDAlZb3VyU3RhdGUxETAPBgNVBAcM
+CFlvdXJDaXR5MRkwFwYDVQQKDBBZb3VyT3JnYW5pemF0aW9uMREwDwYDVQQLDAhZ
+b3VyVW5pdDEcMBoGA1UEAwwTc3JjLWRhdGEtcmVwby5jby51azAeFw0yNDA3MjQx
+MzMxNThaFw0yOTA3MjMxMzMxNThaMIGAMQswCQYDVQQGEwJVSzESMBAGA1UECAwJ
+WW91clN0YXRlMREwDwYDVQQHDAhZb3VyQ2l0eTEZMBcGA1UECgwQWW91ck9yZ2Fu
+aXphdGlvbjERMA8GA1UECwwIWW91clVuaXQxHDAaBgNVBAMME3NyYy1kYXRhLXJl
+cG8uY28udWswggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCrprA6xONC
+kzmcEWzxhGqPeWVHJ5ax5k3PlqZp4saN+w5ho9RzK32iXJy66T0hm5wU5gZnm8X/
++5YVUyo/Zako4gMEzesyN+WSSOPsXBOev1JHz+ekDZTm7RXjDnrlLdhHNhSkY+9F
+y3KZkA4XEy2Jr9HlBmzd9V2rvdr6Y5H15CTyqLtajFqj61LHxWxrBg9zz+5pZYwF
+8l+X8tEaGVuRms0w4IxRpnpP9FbT/ezxGLzaEtgEkTgt2wB9HIkVyAtbiq5qOuES
+pXvgf/EHKXs+bqTwZLtd2nxrUFpF3Zbb9WHwI7hlfJ4jvwVmjpX644E8WyDz/Ivx
++m/3aR6LLvZdAgMBAAGjUzBRMB0GA1UdDgQWBBSTpCFqpJVCtIrxUEV72Ft274wX
+mTAfBgNVHSMEGDAWgBSTpCFqpJVCtIrxUEV72Ft274wXmTAPBgNVHRMBAf8EBTAD
+AQH/MA0GCSqGSIb3DQEBCwUAA4IBAQA0EMO8tYj//CjKgcklMBwuzyXKi2XP8HqS
+18eygdWOywAHnUkP8ha7+CYkW1wyRx63ehWWamAi076KRX/whfNEwyU8o5BIwMko
+qiaiIPSvA29PAuMB+vG5Vzr4NeWQo/YqEGm9zYHRn5F7RwlezjD80llVLn+FWP+a
+meIFHubIL3NGM+J8kUN9m2+/q1L5BkWG6UMM31Br8DastJLMFrzrNICVcWnbtP5B
+ae0aUacJ0r3LakSN3dqjpwjZ/V8nYe3BzRR2Z39HxPTbCWoagwE/Q2SVQFIr1F53
+5rS9RWLKYXIwebRdJZlQEfJ9QUUJ4n6BHT6wvmyEFqvJjFYm83AJ
+-----END CERTIFICATE-----
