+++ scripts/state.yml
@@ -1,0 +1,3 @@
+bookmarks:
+  blank_timestamp:
+    last_record: 14-Dec-2021 21:59
